Rename BoreholeGeometryFormat to BoreholeGeometry
